+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+- `database::status64` wrapping for `sqlite3_db_status64`
+
 ## [1.7] - 2026-05-26
 
 ### Fixed